refactor: move `ArticlePreview` from `domain_entities` to `aggregations` module
