+++ src/domain/aggregations/mod.rs
@@ -1,0 +1,1 @@
+pub mod article_preview;
