+++ src/tests/repositories/article_repository.rs
@@ -4,8 +4,8 @@ use uuid::Uuid;
 
 use super::LocalDb;
 use crate::core::pagination::PaginationParameters;
+use crate::domain::aggregations::article_preview::{ArticlePreview, ArticlePreviewAuthor};
 use crate::domain::domain_entities::article::Article;
-use crate::domain::domain_entities::article_preview::{ArticlePreview, ArticlePreviewAuthor};
 use crate::domain::domain_entities::slug::Slug;
 use crate::domain::domain_entities::user::User;
 use crate::domain::repositories::article_repository::{
